human friendly dao filter query, WagmiWrapper
