+++ .env.template
@@ -20,6 +20,7 @@ AXIOM_DATASET="andreivdev"
 WEB_URL="http://localhost:3000"
 NEXTAUTH_SECRET="ZOMRN3yDm9Bnx0MiQLgrA3o76scTar9HuS4FHjkHhcw="
 NEXTAUTH_URL="http://localhost:3000"
+NEXT_PUBLIC_PROVIDER_URL=
 
 
 # Eth node providers
