+++ apps/senate/package.json
@@ -27,14 +27,9 @@
         "@rainbow-me/rainbowkit-siwe-next-auth": "^0.1.5",
         "@senate/config": "*",
         "@senate/database": "*",
-        "@trpc/client": "10.9.0",
-        "@trpc/next": "10.9.0",
-        "@trpc/react-query": "10.9.0",
-        "@trpc/server": "10.9.0",
         "core-js": "^3.27.2",
         "dayjs": "^1.11.7",
         "ethers": "^5.7.2",
-        "fast-average-color": "^9.3.0",
         "fast-average-color-node": "^2.6.0",
         "framer-motion": "^8.5.4",
         "next": "13.1.6",
@@ -44,7 +39,6 @@
         "react-icons": "^4.7.1",
         "react-query": "^3.39.3",
         "siwe": "^2.1.3",
-        "superjson": "^1.12.2",
         "swr": "^2.0.3",
         "wagmi": "^0.11.0",
         "zod": "^3.20.2"
