+++ apps/senate/src/app/components/csr/RainbowConnect.tsx
@@ -1,46 +1,10 @@
 'use client'
 
-import {
-    ConnectButton,
-    RainbowKitProvider,
-    darkTheme,
-    getDefaultWallets
-} from '@rainbow-me/rainbowkit'
-import { SessionProvider, useSession } from 'next-auth/react'
-import type { GetSiweMessageOptions } from '@rainbow-me/rainbowkit-siwe-next-auth'
-import { RainbowKitSiweNextAuthProvider } from '@rainbow-me/rainbowkit-siwe-next-auth'
-import { configureChains, mainnet, createClient, WagmiConfig } from 'wagmi'
-import { jsonRpcProvider } from 'wagmi/providers/jsonRpc'
-import { publicProvider } from 'wagmi/providers/public'
+import { ConnectButton } from '@rainbow-me/rainbowkit'
+import { useSession } from 'next-auth/react'
 import { useEffect } from 'react'
 import { useRouter } from 'next/navigation'
-
-const { chains, provider } = configureChains(
-    [mainnet],
-    [
-        jsonRpcProvider({
-            rpc: () => ({
-                http: process.env.NEXT_PUBLIC_PROVIDER_URL ?? 'missing_key'
-            })
-        }),
-        publicProvider()
-    ]
-)
-
-const { connectors } = getDefaultWallets({
-    appName: 'Senate',
-    chains
-})
-
-const wagmiClient = createClient({
-    autoConnect: true,
-    connectors,
-    provider
-})
-
-const getSiweMessageOptions: GetSiweMessageOptions = () => ({
-    statement: 'Sign in to Senate'
-})
+import { WagmiWrapper } from './WagmiWrapper'
 
 const RainbowConnectButton = () => {
     // const session = useSession()
@@ -71,26 +35,8 @@ const RainbowConnectButton = () => {
 
 export const RainbowConnect = () => {
     return (
-        <>
-            <WagmiConfig client={wagmiClient}>
-                <SessionProvider refetchInterval={60}>
-                    <RainbowKitSiweNextAuthProvider
-                        getSiweMessageOptions={getSiweMessageOptions}
-                    >
-                        <RainbowKitProvider
-                            chains={chains}
-                            theme={darkTheme({
-                                accentColor: '#262626',
-                                accentColorForeground: 'white',
-                                borderRadius: 'medium',
-                                overlayBlur: 'small'
-                            })}
-                        >
-                            <RainbowConnectButton />
-                        </RainbowKitProvider>
-                    </RainbowKitSiweNextAuthProvider>
-                </SessionProvider>
-            </WagmiConfig>
-        </>
+        <WagmiWrapper>
+            <RainbowConnectButton />
+        </WagmiWrapper>
     )
 }
