+++ apps/senate/src/app/daos/page.tsx
@@ -2,24 +2,28 @@ import { Suspense } from 'react'
 import SubscribedDAOs from './subscribedDAOs/page'
 import UnsubscribedDAOs from './unsubscribedDAOs/page'
 import Loading from './subscribedDAOs/loading'
+import { getServerSession } from 'next-auth'
 
 export default async function Home() {
+    const session = await getServerSession()
     return (
         <main className='text-white'>
             <div className='flex grow flex-col p-5'>
                 <div className='w-full'>
-                    <div className='w-full p-10'>
-                        <p className='mb-4 w-full text-[36px] font-medium text-white'>
-                            Your DAOs
-                        </p>
+                    {session && (
+                        <div className='w-full p-10'>
+                            <p className='mb-4 w-full text-[36px] font-medium text-white'>
+                                Your DAOs
+                            </p>
 
-                        <div className='w-full'>
-                            <Suspense fallback={<Loading />}>
-                                {/* @ts-expect-error Server Component */}
-                                <SubscribedDAOs />
-                            </Suspense>
+                            <div className='w-full'>
+                                <Suspense fallback={<Loading />}>
+                                    {/* @ts-expect-error Server Component */}
+                                    <SubscribedDAOs />
+                                </Suspense>
+                            </div>
                         </div>
-                    </div>
+                    )}
 
                     <div className='p-10'>
                         <p className='mb-4 w-full text-[36px] font-medium text-white'>
