+++ apps/senate/src/app/proposals/active/components/csr/Filters.tsx
@@ -83,7 +83,7 @@ export const Filters = (props: {
                         </option>
                         {props.subscriptions.map((sub) => {
                             return (
-                                <option key={sub.id} value={sub.id}>
+                                <option key={sub.name} value={sub.name}>
                                     {sub.name}
                                 </option>
                             )
